Created a parser utility file.
